+++ src/lib.rs
@@ -7,7 +7,15 @@
 //!
 //! [interp1d::Interp1D] provides functionality to interpolate _n_-dimensional
 //! arrays along the first axis. The documentation of the different
-//! interpolation methods provides examples.
+//! interpolation methods provide examples.
+//!
+//! # Custom interpolation strategy
+//! This crate defines the traits [`StrategyBuilder`](interp1d::StrategyBuilder)
+//! and [`Strategy`](interp1d::Strategy) that allow implementation of user
+//! defined interpolation algorithms.
+//! see the `custom_strategy.rs` example.
+//!
+
 use thiserror::Error;
 
 pub mod interp1d;